+++ src/ios/BrazePlugin.h
@@ -6,8 +6,15 @@
 @property Braze *braze;
 @property NSMutableArray<BRZCancellable *> *subscriptions;
 
+/// Static Braze instance provided as a convenience method for compatibility.
+///
+/// Accessing this property before initializing the Braze plugin will return a `nil` value.
++ (Braze *)braze;
+
 /*-------Braze-------*/
 - (void)changeUser:(CDVInvokedUrlCommand *)command;
+- (void)setSdkAuthenticationSignature:(CDVInvokedUrlCommand *)command;
+- (void)subscribeToSdkAuthenticationFailures:(CDVInvokedUrlCommand *)command;
 - (void)logCustomEvent:(CDVInvokedUrlCommand *)command;
 - (void)logPurchase:(CDVInvokedUrlCommand *)command;
 - (void)disableSdk:(CDVInvokedUrlCommand *)command;
@@ -72,5 +79,6 @@
 - (void)getFeatureFlagBooleanProperty:(CDVInvokedUrlCommand *)command;
 - (void)getFeatureFlagStringProperty:(CDVInvokedUrlCommand *)command;
 - (void)getFeatureFlagNumberProperty:(CDVInvokedUrlCommand *)command;
+- (void)logFeatureFlagImpression:(CDVInvokedUrlCommand *)command;
 
 @end